Split gathering and validation of links into two plugins
