+++ website/gatsby-node.js
@@ -260,3 +260,15 @@ function getGitLog(filepath) {
 
   return git().log(logOptions);
 }
+
+// exports.onPostBootstrap = ({ getNodesByType, getNodes }) => {
+//   const allNodes = getNodes();
+//   const fileNodes = getNodesByType("File");
+//   const mdxNodes = getNodesByType("Mdx");
+
+//   fs.writeFileSync("allNodes.json", JSON.stringify(allNodes));
+//   fs.writeFileSync("fileNodes.json", JSON.stringify(fileNodes));
+//   fs.writeFileSync("mdxNodes.json", JSON.stringify(mdxNodes));
+
+//   console.log("bruh");
+// };
