--- website/plugins/detect-broken-links/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "detect-broken-links",
-  "main": "index.js"
-}
