+++ website/plugins/gatsby-remark-gather-links/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "gatsby-remark-gather-links",
+  "main": "index.js"
+}
