+++ website/plugins/shared/index.js
@@ -1,0 +1,5 @@
+function getCacheKey(node) {
+  return `gatsby-remark-gather-links-${node.id}-${node.internal.contentDigest}`;
+}
+
+module.exports = { getCacheKey };
